backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose
openai-whisper
torch
